+++ packages/dynamic-links/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.3.2](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/dynamic-links@7.3.1...@react-native-firebase/dynamic-links@7.3.2) (2020-07-09)
+
+**Note:** Version bump only for package @react-native-firebase/dynamic-links
+
+## [7.3.1](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/dynamic-links@7.3.0...@react-native-firebase/dynamic-links@7.3.1) (2020-07-07)
+
+**Note:** Version bump only for package @react-native-firebase/dynamic-links
+
+# [7.3.0](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/dynamic-links@7.2.4...@react-native-firebase/dynamic-links@7.3.0) (2020-07-07)
+
+### Features
+
+- **android,ios:** upgrade native SDK versions ([#3881](https://github.com/invertase/react-native-firebase/issues/3881)) ([6cb68a8](https://github.com/invertase/react-native-firebase/commit/6cb68a8ea808392fac3a28bdb1a76049c7b52e86))
+
+## [7.2.4](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/dynamic-links@7.2.3...@react-native-firebase/dynamic-links@7.2.4) (2020-07-05)
+
+**Note:** Version bump only for package @react-native-firebase/dynamic-links
+
+## [7.2.3](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/dynamic-links@7.2.2...@react-native-firebase/dynamic-links@7.2.3) (2020-07-05)
+
+### Bug Fixes
+
+- **dynamic-links,ios:** resolveLink() does not always resolve ([#3870](https://github.com/invertase/react-native-firebase/issues/3870)) ([f10658e](https://github.com/invertase/react-native-firebase/commit/f10658e11aaae9a5b1b2566c05a0a62c46a00522))
+
 ## [7.2.2](https://github.com/invertase/react-native-firebase/compare/@react-native-firebase/dynamic-links@7.2.1...@react-native-firebase/dynamic-links@7.2.2) (2020-06-30)
 
 **Note:** Version bump only for package @react-native-firebase/dynamic-links